survey-codification/requirements.txt
Flask==1.0.2
python-docx==0.8.10